Login logic isolated in the services module.
